+++ tests/unit/test_feature_flag_service_without_cache.py
@@ -22,7 +22,7 @@ class TestFeatureFlagServiceWithoutCache(unittest.IsolatedAsyncioTestCase):
         existing_flag = FeatureFlag(
             id=flag_id, name=random_word(), code=random_word(), enabled=True
         )
-        self.mock_repository.get_by_id.return_value = existing_flag
+        self.mock_repository.get_by_code.return_value = existing_flag
 
         await self.service.disable_feature_flag(flag_id)
 
@@ -36,7 +36,7 @@ class TestFeatureFlagServiceWithoutCache(unittest.IsolatedAsyncioTestCase):
         existing_flag = FeatureFlag(
             id=flag_id, name=random_word(), code=random_word(), enabled=False
         )
-        self.mock_repository.get_by_id.return_value = existing_flag
+        self.mock_repository.get_by_code.return_value = existing_flag
 
         await self.service.enable_feature_flag(flag_id)
 
@@ -54,9 +54,15 @@ class TestFeatureFlagServiceWithoutCache(unittest.IsolatedAsyncioTestCase):
         flag_data = FeatureFlag(
             id=str(uuid.uuid4()), name=random_word(), code=random_word()
         )
-        self.mock_repository.get_by_id.return_value = flag_data
-        result = await self.service.get_feature_flag(flag_data.id)
+        self.mock_repository.get_by_code.return_value = flag_data
+        result = await self.service.get_feature_flag_by_code(flag_data.code)
         self.assertEqual(result, flag_data)
+
+    async def test_list_feature_flags(self):
+        limit = 10
+        skip = 2
+        await self.service.list_feature_flags(limit=limit, skip=skip)
+        self.mock_repository.list_feature_flags.assert_called_once_with(limit=limit, skip=skip, entity_class=FeatureFlag)
 
     async def test_update_feature_flag(self):
         flag_data = {"name": random_word(), "code": random_word()}
@@ -64,9 +70,10 @@ class TestFeatureFlagServiceWithoutCache(unittest.IsolatedAsyncioTestCase):
         self.mock_repository.update.assert_called_once()
 
     async def test_delete_feature_flag(self):
-        flag_id = str(uuid.uuid4())
-        await self.service.delete_feature_flag(flag_id)
-        self.mock_repository.delete.assert_called_once_with(entity_id=flag_id)
+        feature_flag = FeatureFlag(id=str(uuid.uuid4()), name=random_word(), code=random_word())
+        self.mock_repository.get_by_code.return_value = feature_flag
+        await self.service.delete_feature_flag(feature_flag.code)
+        self.mock_repository.delete.assert_called_once_with(entity_id=feature_flag.id)
 
 
 if __name__ == "__main__":