+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - First release with CRUD operations for feature flags
 - Use AsyncSession for database connection and Redis Cluster
+- Updating enable / disable, update, get based on the code instead of id
+- List feature flags based on the offset and skip